+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Don't generate types for parameters with an `[FromServices]` annotation (#95)
 - Add default `any`-mapping for `IActionResult` as a return type (#98)
+- Log an error if we're unable to match a route parameter when generating API clients (#86)
 
 ### Changed